n8n-playground/requirements.txt
# n8n API Playground - Python Dependencies
# Core framework and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Redis and caching
redis==5.0.1
aioredis==2.0.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client and requests
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0

# Authentication and security
pyjwt==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
oauthlib==3.2.2
authlib==1.2.1

# Template engine
jinja2==3.1.2
markupsafe==2.1.3
pyyaml==6.0.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0

# Task queue and background jobs
celery==5.3.4
kombu==5.3.4

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# Environment and configuration
python-dotenv==1.0.0
click==8.1.7

# Testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # for testing FastAPI

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocstrings==0.24.0

# Utilities
typing-extensions==4.8.0
rich==13.7.0

# Workflow and template processing
jsonschema==4.20.0
cerberus==1.3.5

# File handling
aiofiles==23.2.1

# Cryptography
cryptography==45.0.3

# Email handling
aiosmtplib==3.0.1

# Metrics and monitoring
psutil==5.9.6

# Development and debugging
ipython==8.17.2
ipdb==0.13.13

# Optional: Machine Learning (for future anomaly detection)
# scikit-learn==1.3.2
# numpy==1.24.4
# pandas==2.1.4

# Optional: Advanced monitoring
# grafana-api==1.0.3
# prometheus-api-client==0.5.3

# Optional: Message queues
# pika==1.3.2  # RabbitMQ
# kafka-python==2.0.2  # Apache Kafka